+++ api/src/main/java/org/accula/api/code/FileFilter.java
@@ -5,6 +5,8 @@ import java.util.function.Predicate;
 /**
  * Interface to filter files by their names.
  * If {@link FileFilter#test} returns false, file won't be processed.
+ * 
+ * @author Vadim Dyachkov
  */
 @FunctionalInterface
 public interface FileFilter extends Predicate<String> {
