+++ api/src/main/java/org/accula/api/code/RepositoryProvider.java
@@ -1,0 +1,11 @@
+package org.accula.api.code;
+
+import org.eclipse.jgit.lib.Repository;
+import reactor.core.publisher.Mono;
+
+/**
+ * @author Vadim Dyachkov
+ */
+public interface RepositoryProvider {
+    Mono<Repository> getRepository(String owner, String repo);
+}
