+++ api/src/main/java/org/accula/api/code/package-info.java
@@ -1,0 +1,4 @@
+@NonNullApi
+package org.accula.api.code;
+
+import org.springframework.lang.NonNullApi;
