+++ api/src/main/java/org/accula/api/config/CodeConfig.java
@@ -1,19 +1,21 @@
 package org.accula.api.config;
 
+import org.accula.api.code.RepositoryProvider;
 import org.accula.api.code.RepositoryManager;
-import org.accula.api.code.RepositoryManagerImpl;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 import java.io.File;
 
+/**
+ * @author Vadim Dyachkov
+ */
 @Configuration
-public class CodeConfiguration {
-
+public class CodeConfig {
     @Bean
-    public RepositoryManager repositoryManager(@Value("${accula.reposPath}") final String reposPath) {
+    public RepositoryProvider repositoryManager(@Value("${accula.reposPath}") final String reposPath) {
         final File root = new File(reposPath);
-        return new RepositoryManagerImpl(root);
+        return new RepositoryManager(root);
     }
 }
