+++ api/src/main/resources/application.properties
@@ -1,7 +1,6 @@
 PORT=8080
 API_URL=http://localhost:8080
 WEB_URL=http://localhost:3000
-CODE_URL=http://localhost:8081
 DB_HOST=localhost
 DB_PORT=5432
 DB_USER=postgres
